Made(BALL_4000FLX) the fallback device
A very sad hack to work around the Spaceball 5000 FLX-A.

It responds with a hybrid SBall/Magellan ID which results in a NULL modelId and no recorded device.version_string.  The NULL is then to setupDevice with the original code.

Internally the 5000 FLX-A operates indetically to a Spaceball 4000FLX. This addition makes that the default state should a NULL modelId be detected.  All other known devices still detect properly.  Unidentified devices will non-NULL modelID will still display their device.version_string properly for troubleshooting.